test: separate test helpers with import cycles
Separate test helpers with import cycles by moving the helpers to the
internal/testarch and internal/testdistro packages. This now allows using
the helpers in other packages, specifically in the osbuild package.
