+++ internal/test/helpers.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"fmt"
 	"io"
 	"net"
 	"net/http"
@@ -18,8 +17,6 @@ import (
 	"github.com/BurntSushi/toml"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/uuid"
-	"github.com/osbuild/images/pkg/arch"
-	"github.com/osbuild/images/pkg/distro"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -196,17 +193,6 @@ func Ignore(what string) cmp.Option {
 	return cmp.FilterPath(func(p cmp.Path) bool { return p.String() == what }, cmp.Ignore())
 }
 
-// CompareImageType considers two image type objects equal if and only if the names of their distro/arch/imagetype
-// are. The thinking is that the objects are static, and resolving by these three keys should always give equivalent
-// objects. Whether we actually have object equality, is an implementation detail, so we don't want to rely on that.
-func CompareImageTypes() cmp.Option {
-	return cmp.Comparer(func(x, y distro.ImageType) bool {
-		return x.Name() == y.Name() &&
-			x.Arch().Name() == y.Arch().Name() &&
-			x.Arch().Distro().Name() == y.Arch().Distro().Name()
-	})
-}
-
 // Create a temporary repository
 func SetUpTemporaryRepository() (string, error) {
 	dir, err := os.MkdirTemp("/tmp", "osbuild-composer-test-")
@@ -231,20 +217,4 @@ func SetUpTemporaryRepository() (string, error) {
 // Remove the temporary repository
 func TearDownTemporaryRepository(dir string) error {
 	return os.RemoveAll(dir)
-}
-
-// GenerateCIArtifactName generates a new identifier for CI artifacts which is based
-// on environment variables specified by Jenkins
-// note: in case of migration to sth else like Github Actions, change it to whatever variables GH Action provides
-func GenerateCIArtifactName(prefix string) (string, error) {
-	distroCode := os.Getenv("DISTRO_CODE")
-	branchName := os.Getenv("BRANCH_NAME")
-	buildId := os.Getenv("BUILD_ID")
-	if branchName == "" || buildId == "" || distroCode == "" {
-		return "", fmt.Errorf("The environment variables must specify BRANCH_NAME, BUILD_ID, and DISTRO_CODE")
-	}
-
-	arch := arch.Current().String()
-
-	return fmt.Sprintf("%s%s-%s-%s-%s", prefix, distroCode, arch, branchName, buildId), nil
 }
